There is a file "修改內容.txt" for you to check.

+++ proyekDuaOOP/dice.h
@@ -10,7 +10,7 @@ class Dice {
 public:
 	Dice();
 	std::vector<char> result;
-	double max_movement_points;
+	int max_movement_points;
 	double movement_points = 0.0;
 	double success_percent;
 	void action(Stat& status, int used_focus);
